Change the size of the column.
MySQL have a hard limit on the size of the index for InnoDB (767
bytes).  Since the table use the utf8 charset, that let us only 255
characters for our keys.  By reducing the size of the column
`distribution`, `author` and `version`, we can fit in this constraint.

In the current database, the longest version number is 18 characters
long, the longest distribution name is 122 characters, and the length
for the authors's ID is 50.